python/encrypt/requirements.txt
pycryptodomex==3.8.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
